\name{addfour}
\alias{addfour}

\title{addfour}

\description{This function adds 4 to the input variable}

\usage{
addfour(y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{input variable}
	}

\details{Not needed}

\value{numeric value is returned}

\references{Not available}
\author{Vikrant Sihag}

\note{No notes needed}


\examples{
y<-4

print(addfour(y))
}

\keyword{addnumbers}