\name{addtwo}
\alias{addtwo}

\title{addtwo}

\description{This function adds 2 to the input variable}

\usage{
addtwo(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Input variable}
	}

\details{Not needed}

\value{Numeric value is returned}

\references{Not available}
\author{Vikrant Sihag}

\note{No notes needed}


\examples{
x<-2

print(addtwo(x))
}

\keyword{addnumbers}

