\name{vikrantpkg-package}
\alias{vikrantpkg-package}
\alias{vikrantpkg}
\docType{package}
\title{This package add numbers}

\description{This package consists of two functions addtwo and addfour}

\details{
\tabular{ll}{
Package: \tab vikrantpkg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-10\cr
License: \tab GPL 2\cr
}

}
\author{
Vikrant Sihag

Maintainer:Vikrant Sihag <vikrantsihag@gmail.com>

}

\references{Not needed}


\keyword{addnumbers}

\examples{
test1<-addtwo(2)
test2<-addfour(4)
}
